requirements.txt
fastapi
uvicorn
celery
flower
httpx
pika
sqlalchemy
ruff
alembic
pydantic-settings
python-dotenv
psycopg[binary]
sqlmodel
aiohttp
aio-pika
asyncpg